Unify auth api types
